Allow setting of primary and/or non-supplementary read selection form XML for both Picard and Biobambam. Fix sorting expression. Refactorings and smaller bugfixes.

+++ resources/analysisTools/bam2fastq/bam2Fastq.sh
@@ -14,6 +14,7 @@
 # PICARD_OPTIONS: Additional options for picard.
 # JAVA_BINARY: Java binary name/path.
 # JAVA_OPTIONS: Defaults to JAVA_OPTS.
+# excludedReadFlags: space delimited list flags for reads to exclude during processing of: secondary, supplementary
 # outputPerReadGroup: Write separate FASTQs for each read group into $outputDir/$basename/ directory. Otherwise
 #                     create $outputDir/${basename}_r{1,2}.fastq{.gz,} files.
 # writeUnpairedFastq: Additionally write a FASTQ with unpaired reads. Otherwise no such file is written.
@@ -53,7 +54,103 @@ fastqForGroupIndex() {
     echo "${files[0]}"
 }
 
-processPairedEndWithReadGroups() {
+biobambamCompressIntermediateFastqs() {
+    if [[ "$compressIntermediateFastqs" == true ]]; then
+        echo 1
+    else
+        echo 0
+    fi
+}
+
+checkExclusions() {
+    declare -a flagList=($@)
+    for flag in "${flagList[@]}"; do
+       if [[ "$flag" != "secondary" && "$flag" != "supplementary" ]]; then
+          throw 20 "Cannot set '$flag' flag."
+       fi
+    done
+}
+
+bamtofastqExclusions() {
+    declare -a _excludedReadFlags=("${excludedReadFlags[@]:-}")
+    checkExclusions "${_excludedReadFlags[@]}"
+    stringJoin "," "${_excludedReadFlags[@]^^}"
+}
+
+processPairedEndWithReadGroupsBiobambam() {
+    local baseName=$(basename "$FILENAME_BAM" .bam)
+
+    ## Write all read-group FASTQs into a directory.
+    local tmpReadGroupDir="$outputAnalysisBaseDirectory/${baseName}_bam2fastq_temp"
+    registerTmpFile "$tmpReadGroupDir"
+    mkdir -p "$tmpReadGroupDir" || throw 1 "Could not create output directory '$tmpReadGroupDir'"
+
+    ## Only process the non-supplementary (-F 0x800), primary (-F 0x100) alignments. BWA flags chimeric alignments as supplementary while the
+    ## full-length reads are exactly the ones not flagged supplementary. See http://seqanswers.com/forums/showthread.php?t=40239
+    ##
+    ## Biobambam bam2fastq (2.0.87)
+    ##
+    ## * takes care of restricting to non-supplementary, primary reads
+    ## * requires collation to produce files split by read-groups
+    ##
+    "$BIOBAMBAM_BAM2FASTQ_BINARY" \
+        filename="$FILENAME_BAM" \
+        T="$tmpReadGroupDir/$baseName.bamtofastq_tmp" \
+        outputperreadgroup=1 \
+        outputperreadgrouprgsm=1 \
+        outputdir="$tmpReadGroupDir" \
+        collate=1 \
+        colsbs=268435456 \
+        colhlog=19 \
+        gz=$(biobambamCompressIntermediateFastqs) \
+        outputperreadgroupsuffixF=_R1."$FASTQ_SUFFIX" \
+        outputperreadgroupsuffixF2=_R2."$FASTQ_SUFFIX" \
+        outputperreadgroupsuffixO=_U1."$FASTQ_SUFFIX" \
+        outputperreadgroupsuffixO2=_U2."$FASTQ_SUFFIX" \
+        outputperreadgroupsuffixS=_S."$FASTQ_SUFFIX" \
+        outputperreadgrouprgsm=0 \
+        exclude=$(bamtofastqExclusions)
+
+    # Reads without group are assigned to 'default' group.
+
+    ## Now make sure that the output files are renamed correctly.
+    for readGroup in "${readGroups[@]}"; do
+        ## TODO Also rescue U1 U2 S files.
+        for read in R1 R2; do
+            local srcName="$tmpReadGroupDir/${readGroup}_${read}.$FASTQ_SUFFIX"
+            local fgindex="${readGroup}_${read}"
+            local tgtName=$(fastqForGroupIndex "$fgindex")
+            if [[ -f "$srcName" ]]; then
+                ## The file for the "default" read group is only produced if reads exist that are not assigned to a group. Everything produced
+                ## can be moved, though.
+                mv "$srcName" "$tgtName" || throw 10 "Could not move file '$srcName' to '$tgtName'"
+            else
+                ## Furthermore, for read groups in the SAM header without reads, no FASTQ is produced. Such read group produce problems
+                ## downstream. So here we create an empty dummy FASTQ. Note that this may include the 'default' read group that Roddy adds
+                ## See "bamListReadGroups.sh".
+                cat /dev/null | gzip -c - > "$tgtName"
+            fi
+        done
+    done
+}
+
+samtoolsExclusions() {
+    declare -a excludedReadFlagsArray=("${excludedReadFlags[@]:-}")
+    checkExclusions "${excludedReadFlagsArray[@]}"
+    declare exclusionFlag=0
+    declare exclusionsString=$(stringJoin "," "${excludedReadFlagsArray[@]}")
+    if (echo "$exclusionsString" | grep -wq "supplementary"); then
+        let exclusionFlag=($exclusionFlag + 2048)
+    fi
+    if (echo "$exclusionsString" | grep -wq "secondary"); then
+        let exclusionFlag=($exclusionFlag + 256)
+    fi
+    if [[ $exclusionFlag -gt 0 ]]; then
+        echo "-F $exclusionFlag"
+    fi
+}
+
+processPairedEndWithReadGroupsPicard() {
     local baseName=$(basename "$FILENAME_BAM" .bam)
 
     ## Write all read-group FASTQs into a directory.
@@ -63,25 +160,33 @@ processPairedEndWithReadGroups() {
 
     local PICARD_OPTIONS="$PICARD_OPTIONS COMPRESS_OUTPUTS_PER_RG=$compressIntermediateFastqs OUTPUT_PER_RG=true RG_TAG=${readGroupTag:-id} OUTPUT_DIR=${tmpReadGroupDir}"
     local JAVA_OPTIONS="${JAVA_OPTIONS:-$JAVA_OPTS}"
-    ## Only process the non-supplementary reads (-F 0x800). BWA flags all alternative alignments as supplementary while the full-length
-    ## reads are exactly the ones not flagged supplementary.
-    "$SAMTOOLS_BINARY" view -u -F 0x800 "$FILENAME_BAM" \
+    ## Only process the non-supplementary (-F 0x800), primary (-F 0x100) alignments. BWA flags chimeric alignments as supplementary while the
+    ## full-length reads are exactly the ones not flagged supplementary. See http://seqanswers.com/forums/showthread.php?t=40239.
+    "$SAMTOOLS_BINARY" view -u $(samtoolsExclusions) "$FILENAME_BAM" \
         | "$PICARD_BINARY" $JAVA_OPTIONS SamToFastq $PICARD_OPTIONS INPUT=/dev/stdin
 
     ## Now make sure that the output files of Picard are renamed correctly.
-    ## TODO Get the read groups from the java code.
     for readGroup in "${readGroups[@]}"; do
         for read in 1 2; do
             local srcName="$tmpReadGroupDir/${readGroup}_${read}.$FASTQ_SUFFIX"
             local fgindex="${readGroup}_R${read}"
             local tgtName=$(fastqForGroupIndex "$fgindex")
-            mv "$srcName" "$tgtName" || throw 10 "Could not move file '$srcName' to '$tgtName'"
+            if [[ -f "$srcName" ]]; then
+                ## The file for the "default" read group is only produced if reads exist that are not assigned to a group. Everything produced
+                ## can be moved, though.
+                mv "$srcName" "$tgtName" || throw 10 "Could not move file '$srcName' to '$tgtName'"
+            else
+                ## Furthermore, for read groups in the SAM header without reads, no FASTQ is produced. Such read group produce problems
+                ## downstream. So here we create an empty dummy FASTQ. Note that this may include the 'default' read group that Roddy adds
+                ## See "bamListReadGroups.sh".
+                cat /dev/null | gzip -c - > "$tgtName"
+            fi
         done
     done
 }
 
 
-processPairedEndWithoutReadGroups() {
+processPairedEndWithoutReadGroupsPicard() {
     local baseName=$(basename "$FILENAME_BAM" .bam)
 
     ## Extract the output directory from the first lane file path entered.
@@ -104,7 +209,7 @@ processPairedEndWithoutReadGroups() {
     ## Only process the non-supplementary reads (-F 0x800). BWA flags all alternative alignments as supplementary while the full-length
     ## reads are exactly the ones not flagged supplementary.
     local JAVA_OPTIONS="${JAVA_OPTIONS:-$JAVA_OPTS}"
-    "$SAMTOOLS_BINARY" view -u -F 0x800 "$FILENAME_BAM" \
+    "$SAMTOOLS_BINARY" view -u $(samtoolsExclusions) "$FILENAME_BAM" \
         | "$PICARD_BINARY" $JAVA_OPTIONS SamToFastq $PICARD_OPTIONS INPUT=/dev/stdin
 
     ## Now make sure that the output files of Picard are renamed correctly.
@@ -115,11 +220,11 @@ processPairedEndWithoutReadGroups() {
     fi
 }
 
-processSingleEndWithReadGroups() {
+processSingleEndWithReadGroupsPicard() {
     throw 1 "processSingleEndWithReadGroups not implemented"
 }
 
-processSingleEndWithoutReadGroups() {
+processSingleEndWithoutReadGroupsPicard() {
     throw 1 "processSingleEndWithoutReadGroups not implemented"
 }
 
@@ -140,15 +245,21 @@ main() {
 
     if [[ "${pairedEnd:-true}" == true ]]; then
         if [[ "${outputPerReadGroup:-true}" == true ]]; then
-            processPairedEndWithReadGroups
+            if [[ "$converter" == "picard" ]]; then
+                processPairedEndWithReadGroupsPicard
+            elif [[ "$converter" == "biobambam" ]]; then
+                processPairedEndWithReadGroupsBiobambam
+            else
+                throw 10 "Unknown bam-to-fastq converter: '$converter'"
+            fi
         else
-            processPairedEndWithoutReadGroups
+            processPairedEndWithoutReadGroupsPicard
         fi
     else
         if [[ "${outputPerReadGroup:-true}" == true ]]; then
-            processSingleEndWithReadGroups
+            processSingleEndWithReadGroupsPicard
         else
-            processSingleEndWithoutReadGroups
+            processSingleEndWithoutReadGroupsPicard
         fi
     fi
